requirements.txt
boto3
fastapi
google-api-python-client
gradio
joblib
mlflow
mlflow_skinny
mlflow_tracing
numpy
optuna
pandas
pydantic
pydantic-settings
scikit-learn==1.6.1 
sqlalchemy
uvicorn
python-dotenv
psycopg2-binary